backend/requirements.txt
fastapi
uvicorn
easyocr
pillow